Move astype, constructor tests
