+++ python/cudf/cudf/tests/test_dataframe.py
@@ -6,7 +6,6 @@ import decimal
 import io
 import operator
 import re
-import string
 import textwrap
 import warnings
 from contextlib import contextmanager
@@ -862,33 +861,6 @@ def test_dataframe_shape_empty():
     assert pdf.shape == gdf.shape
 
 
-@pytest.mark.parametrize("num_cols", [1, 3])
-@pytest.mark.parametrize("num_rows", [1, 5])
-def test_dataframe_transpose_category(num_cols, num_rows):
-    pdf = pd.DataFrame()
-
-    for i in range(num_cols):
-        colname = string.ascii_lowercase[i]
-        data = pd.Series(list(string.ascii_lowercase), dtype="category")
-        data = data.sample(num_rows, replace=True).reset_index(drop=True)
-        pdf[colname] = data
-
-    gdf = cudf.DataFrame.from_pandas(pdf)
-
-    got_function = gdf.transpose()
-    got_property = gdf.T
-
-    expect = pdf.transpose()
-
-    assert_eq(expect, got_function.to_pandas())
-    assert_eq(expect, got_property.to_pandas())
-
-
-def test_generated_column():
-    gdf = cudf.DataFrame({"a": (i for i in range(5))})
-    assert len(gdf) == 5
-
-
 @pytest.fixture
 def pdf():
     return pd.DataFrame({"x": range(10), "y": range(10)})
@@ -1630,354 +1602,6 @@ def test_as_column_types():
     gds = cudf.Series(cudf.Index(["1", "18", "9"]), dtype="int")
 
     assert_eq(pds, gds)
-
-
-@pytest.mark.parametrize("dtype", ALL_TYPES)
-@pytest.mark.parametrize(
-    "np_dtype,pd_dtype",
-    [
-        tuple(item)
-        for item in cudf.utils.dtypes.np_dtypes_to_pandas_dtypes.items()
-    ],
-)
-def test_series_astype_pandas_nullable(dtype, np_dtype, pd_dtype):
-    source = cudf.Series([0, 1, None], dtype=dtype)
-
-    expect = source.astype(np_dtype)
-    got = source.astype(pd_dtype)
-
-    assert_eq(expect, got)
-
-
-@pytest.mark.parametrize("dtype", NUMERIC_TYPES)
-@pytest.mark.parametrize("as_dtype", NUMERIC_TYPES)
-def test_series_astype_numeric_to_numeric(dtype, as_dtype):
-    psr = pd.Series([1, 2, 4, 3], dtype=dtype)
-    gsr = cudf.from_pandas(psr)
-    assert_eq(psr.astype(as_dtype), gsr.astype(as_dtype))
-
-
-@pytest.mark.parametrize("dtype", NUMERIC_TYPES)
-@pytest.mark.parametrize("as_dtype", NUMERIC_TYPES)
-def test_series_astype_numeric_to_numeric_nulls(dtype, as_dtype):
-    data = [1, 2, None, 3]
-    sr = cudf.Series(data, dtype=dtype)
-    got = sr.astype(as_dtype)
-    expect = cudf.Series([1, 2, None, 3], dtype=as_dtype)
-    assert_eq(expect, got)
-
-
-@pytest.mark.parametrize("dtype", NUMERIC_TYPES)
-@pytest.mark.parametrize(
-    "as_dtype",
-    [
-        "str",
-        "category",
-        "datetime64[s]",
-        "datetime64[ms]",
-        "datetime64[us]",
-        "datetime64[ns]",
-    ],
-)
-def test_series_astype_numeric_to_other(dtype, as_dtype):
-    psr = pd.Series([1, 2, 3], dtype=dtype)
-    gsr = cudf.from_pandas(psr)
-    assert_eq(psr.astype(as_dtype), gsr.astype(as_dtype))
-
-
-@pytest.mark.parametrize(
-    "as_dtype",
-    [
-        "str",
-        "int32",
-        "uint32",
-        "float32",
-        "category",
-        "datetime64[s]",
-        "datetime64[ms]",
-        "datetime64[us]",
-        "datetime64[ns]",
-    ],
-)
-def test_series_astype_string_to_other(as_dtype):
-    if "datetime64" in as_dtype:
-        data = ["2001-01-01", "2002-02-02", "2000-01-05"]
-    else:
-        data = ["1", "2", "3"]
-    psr = pd.Series(data)
-    gsr = cudf.from_pandas(psr)
-    assert_eq(psr.astype(as_dtype), gsr.astype(as_dtype))
-
-
-@pytest.mark.parametrize(
-    "as_dtype",
-    [
-        "category",
-        "datetime64[s]",
-        "datetime64[ms]",
-        "datetime64[us]",
-        "datetime64[ns]",
-    ],
-)
-def test_series_astype_datetime_to_other(as_dtype):
-    data = ["2001-01-01", "2002-02-02", "2001-01-05"]
-    psr = pd.Series(data)
-    gsr = cudf.from_pandas(psr)
-    assert_eq(psr.astype(as_dtype), gsr.astype(as_dtype))
-
-
-@pytest.mark.parametrize(
-    "inp",
-    [
-        ("datetime64[ns]", "2011-01-01 00:00:00.000000000"),
-        ("datetime64[us]", "2011-01-01 00:00:00.000000"),
-        ("datetime64[ms]", "2011-01-01 00:00:00.000"),
-        ("datetime64[s]", "2011-01-01 00:00:00"),
-    ],
-)
-def test_series_astype_datetime_to_string(inp):
-    dtype, expect = inp
-    base_date = "2011-01-01"
-    sr = cudf.Series([base_date], dtype=dtype)
-    got = sr.astype(str)[0]
-    assert expect == got
-
-
-@pytest.mark.parametrize(
-    "as_dtype",
-    [
-        "int32",
-        "uint32",
-        "float32",
-        "category",
-        "datetime64[s]",
-        "datetime64[ms]",
-        "datetime64[us]",
-        "datetime64[ns]",
-        "str",
-    ],
-)
-def test_series_astype_categorical_to_other(as_dtype):
-    if "datetime64" in as_dtype:
-        data = ["2001-01-01", "2002-02-02", "2000-01-05", "2001-01-01"]
-    else:
-        data = [1, 2, 3, 1]
-    psr = pd.Series(data, dtype="category")
-    gsr = cudf.from_pandas(psr)
-    assert_eq(psr.astype(as_dtype), gsr.astype(as_dtype))
-
-
-@pytest.mark.parametrize("ordered", [True, False])
-def test_series_astype_to_categorical_ordered(ordered):
-    psr = pd.Series([1, 2, 3, 1], dtype="category")
-    gsr = cudf.from_pandas(psr)
-
-    ordered_dtype_pd = pd.CategoricalDtype(
-        categories=[1, 2, 3], ordered=ordered
-    )
-    ordered_dtype_gd = cudf.CategoricalDtype.from_pandas(ordered_dtype_pd)
-    assert_eq(
-        psr.astype("int32").astype(ordered_dtype_pd).astype("int32"),
-        gsr.astype("int32").astype(ordered_dtype_gd).astype("int32"),
-    )
-
-
-@pytest.mark.parametrize("ordered", [True, False])
-def test_series_astype_cat_ordered_to_unordered(ordered):
-    pd_dtype = pd.CategoricalDtype(categories=[1, 2, 3], ordered=ordered)
-    pd_to_dtype = pd.CategoricalDtype(
-        categories=[1, 2, 3], ordered=not ordered
-    )
-    gd_dtype = cudf.CategoricalDtype.from_pandas(pd_dtype)
-    gd_to_dtype = cudf.CategoricalDtype.from_pandas(pd_to_dtype)
-
-    psr = pd.Series([1, 2, 3], dtype=pd_dtype)
-    gsr = cudf.Series([1, 2, 3], dtype=gd_dtype)
-
-    expect = psr.astype(pd_to_dtype)
-    got = gsr.astype(gd_to_dtype)
-
-    assert_eq(expect, got)
-
-
-def test_series_astype_null_cases():
-    data = [1, 2, None, 3]
-
-    # numerical to other
-    assert_eq(cudf.Series(data, dtype="str"), cudf.Series(data).astype("str"))
-
-    assert_eq(
-        cudf.Series(data, dtype="category"),
-        cudf.Series(data).astype("category"),
-    )
-
-    assert_eq(
-        cudf.Series(data, dtype="float32"),
-        cudf.Series(data, dtype="int32").astype("float32"),
-    )
-
-    assert_eq(
-        cudf.Series(data, dtype="float32"),
-        cudf.Series(data, dtype="uint32").astype("float32"),
-    )
-
-    assert_eq(
-        cudf.Series(data, dtype="datetime64[ms]"),
-        cudf.Series(data).astype("datetime64[ms]"),
-    )
-
-    # categorical to other
-    assert_eq(
-        cudf.Series(data, dtype="str"),
-        cudf.Series(data, dtype="category").astype("str"),
-    )
-
-    assert_eq(
-        cudf.Series(data, dtype="float32"),
-        cudf.Series(data, dtype="category").astype("float32"),
-    )
-
-    assert_eq(
-        cudf.Series(data, dtype="datetime64[ms]"),
-        cudf.Series(data, dtype="category").astype("datetime64[ms]"),
-    )
-
-    # string to other
-    assert_eq(
-        cudf.Series([1, 2, None, 3], dtype="int32"),
-        cudf.Series(["1", "2", None, "3"]).astype("int32"),
-    )
-
-    assert_eq(
-        cudf.Series(
-            ["2001-01-01", "2001-02-01", None, "2001-03-01"],
-            dtype="datetime64[ms]",
-        ),
-        cudf.Series(["2001-01-01", "2001-02-01", None, "2001-03-01"]).astype(
-            "datetime64[ms]"
-        ),
-    )
-
-    assert_eq(
-        cudf.Series(["a", "b", "c", None], dtype="category").to_pandas(),
-        cudf.Series(["a", "b", "c", None]).astype("category").to_pandas(),
-    )
-
-    # datetime to other
-    data = [
-        "2001-01-01 00:00:00.000000",
-        "2001-02-01 00:00:00.000000",
-        None,
-        "2001-03-01 00:00:00.000000",
-    ]
-    assert_eq(
-        cudf.Series(data),
-        cudf.Series(data, dtype="datetime64[us]").astype("str"),
-    )
-
-    assert_eq(
-        pd.Series(data, dtype="datetime64[ns]").astype("category"),
-        cudf.from_pandas(pd.Series(data, dtype="datetime64[ns]")).astype(
-            "category"
-        ),
-    )
-
-
-def test_series_astype_null_categorical():
-    sr = cudf.Series([None, None, None], dtype="category")
-    expect = cudf.Series([None, None, None], dtype="int32")
-    got = sr.astype("int32")
-    assert_eq(expect, got)
-
-
-@pytest.mark.parametrize(
-    "data",
-    [
-        (
-            pd.Series([3, 3.0]),
-            pd.Series([2.3, 3.9]),
-            pd.Series([1.5, 3.9]),
-            pd.Series([1.0, 2]),
-        ),
-        [
-            pd.Series([3, 3.0]),
-            pd.Series([2.3, 3.9]),
-            pd.Series([1.5, 3.9]),
-            pd.Series([1.0, 2]),
-        ],
-    ],
-)
-def test_create_dataframe_from_list_like(data):
-    pdf = pd.DataFrame(data, index=["count", "mean", "std", "min"])
-    gdf = cudf.DataFrame(data, index=["count", "mean", "std", "min"])
-
-    assert_eq(pdf, gdf)
-
-    pdf = pd.DataFrame(data)
-    gdf = cudf.DataFrame(data)
-
-    assert_eq(pdf, gdf)
-
-
-def test_create_dataframe_column():
-    pdf = pd.DataFrame(columns=["a", "b", "c"], index=["A", "Z", "X"])
-    gdf = cudf.DataFrame(columns=["a", "b", "c"], index=["A", "Z", "X"])
-
-    assert_eq(pdf, gdf)
-
-    pdf = pd.DataFrame(
-        {"a": [1, 2, 3], "b": [2, 3, 5]},
-        columns=["a", "b", "c"],
-        index=["A", "Z", "X"],
-    )
-    gdf = cudf.DataFrame(
-        {"a": [1, 2, 3], "b": [2, 3, 5]},
-        columns=["a", "b", "c"],
-        index=["A", "Z", "X"],
-    )
-
-    assert_eq(pdf, gdf)
-
-
-@pytest.mark.parametrize(
-    "data",
-    [
-        pd.DataFrame(np.eye(2)),
-        cudf.DataFrame(np.eye(2)),
-        np.eye(2),
-        cupy.eye(2),
-        None,
-        [[1, 0], [0, 1]],
-        [cudf.Series([0, 1]), cudf.Series([1, 0])],
-    ],
-)
-@pytest.mark.parametrize(
-    "columns",
-    [None, range(2), pd.RangeIndex(2), cudf.RangeIndex(2)],
-)
-def test_dataframe_columns_returns_rangeindex(data, columns):
-    if data is None and columns is None:
-        pytest.skip(f"{data=} and {columns=} not relevant.")
-    result = cudf.DataFrame(data=data, columns=columns).columns
-    expected = pd.RangeIndex(range(2))
-    assert_eq(result, expected)
-
-
-def test_dataframe_columns_returns_rangeindex_single_col():
-    result = cudf.DataFrame([1, 2, 3]).columns
-    expected = pd.RangeIndex(range(1))
-    assert_eq(result, expected)
-
-
-@pytest.mark.parametrize("dtype", ["int64", "datetime64[ns]", "int8"])
-@pytest.mark.parametrize("idx_data", [[], [1, 2]])
-@pytest.mark.parametrize("data", [None, [], {}])
-def test_dataframe_columns_empty_data_preserves_dtype(dtype, idx_data, data):
-    result = cudf.DataFrame(
-        data, columns=cudf.Index(idx_data, dtype=dtype)
-    ).columns
-    expected = pd.Index(idx_data, dtype=dtype)
-    assert_eq(result, expected)
 
 
 @pytest.mark.parametrize("dtype", ["int64", "datetime64[ns]", "int8"])
@@ -5977,24 +5601,6 @@ def test_dataframe_duplicated(data, subset, keep):
 
     expected = pdf.duplicated(subset=subset, keep=keep)
     actual = gdf.duplicated(subset=subset, keep=keep)
-
-    assert_eq(expected, actual)
-
-
-@pytest.mark.parametrize(
-    "data",
-    [
-        {"col": [{"a": 1.1}, {"a": 2.1}, {"a": 10.0}, {"a": 11.2323}, None]},
-        {"a": [[{"b": 567}], None] * 10},
-        {"a": [decimal.Decimal(10), decimal.Decimal(20), None]},
-    ],
-)
-def test_dataframe_transpose_complex_types(data):
-    gdf = cudf.DataFrame(data)
-    pdf = gdf.to_pandas()
-
-    expected = pdf.T
-    actual = gdf.T
 
     assert_eq(expected, actual)
